pyproject.toml
[project]
name = "gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "streamlit>=1.45.1",
]
